backend/requirements-dev.txt
black==24.8.0
isort==5.13.2
flake8==7.1.1
mypy==1.13.0
mypy-extensions==1.0.0
django-stubs==5.0.2
djangorestframework-stubs==3.14.5
pytest==8.4.2
pytest-django==4.11.1
Pillow==11.1.0
